[TIP] IndicatorFlyout => Flyout
