+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/flyout/components/empty_prompt/empty_prompt.tsx
