+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/flyout/components/fields_table/fields_table.tsx
