+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/flyout/tabs/flyout_overview/components/highlighted_values_table/index.tsx
