+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/flyout/tabs/flyout_table/flyout_table.stories.tsx
@@ -1,0 +1,43 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+import { Story } from '@storybook/react';
+import { CoreStart } from '@kbn/core/public';
+import { createKibanaReactContext } from '@kbn/kibana-react-plugin/public';
+import { mockIndicatorsFiltersContext } from '../../../../../../common/mocks/mock_indicators_filters_context';
+import { mockUiSettingsService } from '../../../../../../common/mocks/mock_kibana_ui_settings_service';
+import { mockKibanaTimelinesService } from '../../../../../../common/mocks/mock_kibana_timelines_service';
+import { generateMockIndicator, Indicator } from '../../../../../../../common/types/indicator';
+import { IndicatorFlyoutTable } from '.';
+import { FiltersContext } from '../../../../contexts';
+
+export default {
+  component: IndicatorFlyoutTable,
+  title: 'IndicatorFlyoutTable',
+};
+
+export const Default: Story<void> = () => {
+  const mockIndicator: Indicator = generateMockIndicator();
+
+  const KibanaReactContext = createKibanaReactContext({
+    uiSettings: mockUiSettingsService(),
+    timelines: mockKibanaTimelinesService,
+  } as unknown as CoreStart);
+
+  return (
+    <KibanaReactContext.Provider>
+      <FiltersContext.Provider value={mockIndicatorsFiltersContext}>
+        <IndicatorFlyoutTable indicator={mockIndicator} />
+      </FiltersContext.Provider>
+    </KibanaReactContext.Provider>
+  );
+};
+
+export const EmptyIndicator: Story<void> = () => {
+  return <IndicatorFlyoutTable indicator={{ fields: {} } as unknown as Indicator} />;
+};
