--- x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_flyout/components/indicator_empty_prompt/indicator_empty_prompt.stories.tsx
@@ -1,24 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import React from 'react';
-import { Story } from '@storybook/react';
-import { StoryProvidersComponent } from '../../../../../../common/mocks/story_providers';
-import { IndicatorEmptyPrompt } from './indicator_empty_prompt';
-
-export default {
-  component: IndicatorEmptyPrompt,
-  title: 'IndicatorEmptyPrompt',
-};
-
-export const Default: Story<void> = () => {
-  return (
-    <StoryProvidersComponent>
-      <IndicatorEmptyPrompt />
-    </StoryProvidersComponent>
-  );
-};
