--- x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_flyout/components/indicator_fields_table/indicator_fields_table.stories.tsx
@@ -1,34 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import React from 'react';
-import { mockIndicatorsFiltersContext } from '../../../../../../common/mocks/mock_indicators_filters_context';
-import { IndicatorFieldsTable } from './indicator_fields_table';
-import { generateMockIndicator } from '../../../../../../../common/types/indicator';
-import { StoryProvidersComponent } from '../../../../../../common/mocks/story_providers';
-import { FiltersContext } from '../../../../contexts';
-
-export default {
-  component: IndicatorFieldsTable,
-  title: 'IndicatorFieldsTable',
-};
-
-export function WithIndicators() {
-  const indicator = generateMockIndicator();
-
-  return (
-    <StoryProvidersComponent>
-      <FiltersContext.Provider value={mockIndicatorsFiltersContext}>
-        <IndicatorFieldsTable
-          fields={['threat.indicator.type']}
-          indicator={indicator}
-          search={false}
-        />
-      </FiltersContext.Provider>
-    </StoryProvidersComponent>
-  );
-}
