--- x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_flyout/tabs/indicators_flyout_json/indicators_flyout_json.stories.tsx
@@ -1,26 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import React from 'react';
-import { Story } from '@storybook/react';
-import { generateMockIndicator, Indicator } from '../../../../../../../common/types/indicator';
-import { IndicatorsFlyoutJson } from './indicators_flyout_json';
-
-export default {
-  component: IndicatorsFlyoutJson,
-  title: 'IndicatorsFlyoutJson',
-};
-
-export const Default: Story<void> = () => {
-  const mockIndicator: Indicator = generateMockIndicator();
-
-  return <IndicatorsFlyoutJson indicator={mockIndicator} />;
-};
-
-export const EmptyIndicator: Story<void> = () => {
-  return <IndicatorsFlyoutJson indicator={{} as unknown as Indicator} />;
-};
