--- x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_flyout/tabs/indicators_flyout_overview/components/block/indicator_block.stories.tsx
@@ -1,31 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import React from 'react';
-import { FiltersContext } from '../../../../../../contexts';
-import { StoryProvidersComponent } from '../../../../../../../../common/mocks/story_providers';
-import { generateMockIndicator } from '../../../../../../../../../common/types/indicator';
-import { IndicatorBlock } from './indicator_block';
-
-export default {
-  component: IndicatorBlock,
-  title: 'IndicatorBlock',
-};
-
-const mockIndicator = generateMockIndicator();
-
-export function Default() {
-  const mockField = 'threat.indicator.ip';
-
-  return (
-    <StoryProvidersComponent>
-      <FiltersContext.Provider value={{} as any}>
-        <IndicatorBlock indicator={mockIndicator} field={mockField} />
-      </FiltersContext.Provider>
-    </StoryProvidersComponent>
-  );
-}
