--- x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_flyout/tabs/indicators_flyout_overview/indicators_flyout_overview.test.tsx
@@ -1,48 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { TestProvidersComponent } from '../../../../../../common/mocks/test_providers';
-import { render, screen } from '@testing-library/react';
-import React from 'react';
-import { generateMockIndicator, Indicator } from '../../../../../../../common/types/indicator';
-import {
-  IndicatorsFlyoutOverview,
-  TI_FLYOUT_OVERVIEW_HIGH_LEVEL_BLOCKS,
-  TI_FLYOUT_OVERVIEW_TABLE,
-} from './indicators_flyout_overview';
-import { EMPTY_PROMPT_TEST_ID } from '../../components/indicator_empty_prompt';
-
-describe('<IndicatorsFlyoutOverview />', () => {
-  describe('invalid indicator', () => {
-    it('should render error message on invalid indicator', () => {
-      render(
-        <TestProvidersComponent>
-          <IndicatorsFlyoutOverview
-            onViewAllFieldsInTable={() => {}}
-            indicator={{ fields: {} } as unknown as Indicator}
-          />
-        </TestProvidersComponent>
-      );
-
-      expect(screen.getByTestId(EMPTY_PROMPT_TEST_ID)).toBeInTheDocument();
-    });
-  });
-
-  it('should render the highlighted blocks and table when valid indicator is passed', () => {
-    render(
-      <TestProvidersComponent>
-        <IndicatorsFlyoutOverview
-          onViewAllFieldsInTable={() => {}}
-          indicator={generateMockIndicator()}
-        />
-      </TestProvidersComponent>
-    );
-
-    expect(screen.queryByTestId(TI_FLYOUT_OVERVIEW_TABLE)).toBeInTheDocument();
-    expect(screen.queryByTestId(TI_FLYOUT_OVERVIEW_HIGH_LEVEL_BLOCKS)).toBeInTheDocument();
-  });
-});
