+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_table/indicators_table.test.tsx
@@ -10,8 +10,8 @@ import React from 'react';
 import { IndicatorsTable, IndicatorsTableProps } from './indicators_table';
 import { TestProvidersComponent } from '../../../../common/mocks/test_providers';
 import { generateMockIndicator, Indicator } from '../../../../../common/types/indicator';
-import { BUTTON_TEST_ID } from '../open_indicator_flyout_button/open_indicator_flyout_button';
-import { TITLE_TEST_ID } from '../indicators_flyout/indicators_flyout';
+import { BUTTON_TEST_ID } from '../open_flyout_button';
+import { TITLE_TEST_ID } from '../flyout';
 import { SecuritySolutionDataViewBase } from '../../../../types';
 
 const stub = () => {};
