+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_table/indicators_table.tsx
@@ -23,7 +23,7 @@ import { Indicator, RawIndicatorFieldId } from '../../../../../common/types/indi
 import { cellRendererFactory } from './cell_renderer';
 import { EmptyState } from '../../../../components/empty_state';
 import { IndicatorsTableContext, IndicatorsTableContextValue } from './context';
-import { IndicatorsFlyout } from '../indicators_flyout/indicators_flyout';
+import { IndicatorFlyout } from '../flyout';
 import { Pagination } from '../../hooks/use_indicators';
 import { useToolbarOptions } from './hooks/use_toolbar_options';
 import { ColumnSettingsValue } from './hooks/use_column_settings';
@@ -131,7 +131,7 @@ export const IndicatorsTable: VFC<IndicatorsTableProps> = ({
   const flyoutFragment = useMemo(
     () =>
       expanded ? (
-        <IndicatorsFlyout indicator={expanded} closeFlyout={() => setExpanded(undefined)} />
+        <IndicatorFlyout indicator={expanded} closeFlyout={() => setExpanded(undefined)} />
       ) : null,
     [expanded]
   );
