+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/open_flyout_button/open_flyout_button.tsx
